Refactor - changed all private class fields to auto-properties except userSelectedList in KeyRandomizer
