+++ DvorakTrainer/DvorakTrainer.csproj
@@ -54,6 +54,7 @@
     <Compile Include="Entities\GameTimer.cs" />
     <Compile Include="Entities\KeyRandomizer.cs" />
     <Compile Include="Entities\Program.cs" />
+    <Compile Include="Entities\StatsScore.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Entities\Statistics.cs" />
     <EmbeddedResource Include="Entities\Form1.resx">
