+++ DvorakTrainer/Entities/Form1.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
 using DvorakTrainer;
+using DvorakTrainer.Entities;
 
 namespace Dvorak
 {
