+++ DvorakTrainer/Entities/StatsScore.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DvorakTrainer.Entities
+{
+    class StatsScore : Statistics
+    {
+
+        public int PointsScore { get; set; }
+        public decimal Accuracy { get; set; }
+
+    }
+}
